server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,49 @@ const AuthorizationError = require('../../exceptions/AuthorizationError');
 const { mapDBToModel } = require('../../utils');
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this._cacheService = cacheService;
+  }
+
+  async getNotes(owner) {
+    try {
+      const result = await this._cacheService.get(`notes:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
+        LEFT JOIN collaborations ON collaborations.note_id = notes.id
+        WHERE notes.owner = $1 OR collaborations.user_id = $1
+        GROUP BY notes.id`,
+        values: [owner],
+      };
+
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      await this._cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult));
+
+      return mappedResult;
+    }
+  }
+
+  async getNoteById(id) {
+    const query = {
+      text: `SELECT notes.*, users.username
+      FROM notes
+      LEFT JOIN users ON users.id = notes.owner
+      WHERE notes.id = $1`,
+      values: [id],
+    };
+    const result = await this._pool.query(query);
+
+    if (!result.rows.length) {
+      throw new NotFoundError('Catatan tidak ditemukan');
+    }
+
+    return result.rows.map(mapDBToModel)[0];
   }
 
   async addNote({
@@ -29,42 +69,14 @@ class NotesService {
       throw new InvariantError('Catatan gagal ditambahkan');
     }
 
+    await this._cacheService.delete(`notes:${owner}`);
     return result.rows[0].id;
-  }
-
-  async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-    LEFT JOIN collaborations ON collaborations.note_id = notes.id
-    WHERE notes.owner = $1 OR collaborations.user_id = $1
-    GROUP BY notes.id`,
-      values: [owner],
-    };
-    const result = await this._pool.query(query);
-    return result.rows.map(mapDBToModel);
-  }
-
-  async getNoteById(id) {
-    const query = {
-      text: `SELECT notes.*, users.username
-      FROM notes
-      LEFT JOIN users ON users.id = notes.owner
-      WHERE notes.id = $1`,
-      values: [id],
-    };
-    const result = await this._pool.query(query);
-
-    if (!result.rows.length) {
-      throw new NotFoundError('Catatan tidak ditemukan');
-    }
-
-    return result.rows.map(mapDBToModel)[0];
   }
 
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
 
@@ -73,45 +85,56 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
-
+ 
     const result = await this._pool.query(query);
-
+ 
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
     }
+ 
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
   }
-
+ 
   async verifyNoteOwner(id, owner) {
     const query = {
       text: 'SELECT * FROM notes WHERE id = $1',
       values: [id],
     };
+
     const result = await this._pool.query(query);
+
     if (!result.rows.length) {
       throw new NotFoundError('Catatan tidak ditemukan');
     }
+
     const note = result.rows[0];
+
     if (note.owner !== owner) {
       throw new AuthorizationError('Anda tidak berhak mengakses resource ini');
     }
   }
 
-  async verifyNoteAccess(noteId, userId) {
+  async verifyNoteAccess(id, userId) {
     try {
-      await this.verifyNoteOwner(noteId, userId);
+      await this.verifyNoteOwner(id, userId);
     } catch (error) {
       if (error instanceof NotFoundError) {
         throw error;
       }
+
       try {
-        await this._collaborationService.verifyCollaborator(noteId, userId);
+        await this._collaborationService.verifyCollaborator(id, userId);
       } catch {
         throw error;
       }
